requirements.txt
isla-solver
setuptools

mypy
